+++ server/scheme/scheme.controller.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const adminIdentity = require('../infrastructure/middlewares/adminIdentity');
+const identity = require('../infrastructure/middlewares/identity');
 const SchemeService = require('./scheme.service');
 
 const get = async (req, res, next) => {
@@ -14,6 +15,9 @@ const get = async (req, res, next) => {
 }
 
 const create = async (req, res, next) => {
+  if (!req.user || (!req.user.isAdmin && !req.user.isTournamentAdmin))
+    return next({ name: 'DomainActionError', error: 'notEnoughPermissions' }, req, res, null);
+
   try {
     await SchemeService.create(req.body);
     return res.json({});
@@ -24,6 +28,9 @@ const create = async (req, res, next) => {
 }
 
 const update = async (req, res, next) => {
+  if (!req.user || (!req.user.isAdmin && !req.user.isTournamentAdmin))
+    return next({ name: 'DomainActionError', error: 'notEnoughPermissions' }, req, res, null);
+
   try {
     const scheme = await SchemeService.get(req.params.id);
     await SchemeService.update(scheme, req.body);
@@ -35,6 +42,9 @@ const update = async (req, res, next) => {
 }
 
 const remove = async (req, res, next) => {
+  if (!req.user || (!req.user.isAdmin && !req.user.isTournamentAdmin))
+    return next({ name: 'DomainActionError', error: 'notEnoughPermissions' }, req, res, null);
+
   try {
     await SchemeService.delete(req.params.id);
     return res.json({});
@@ -55,6 +65,9 @@ const include = async (req, res, next) => {
 }
 
 const drawBracket = async (req, res, next) => {
+  if (!req.user || (!req.user.isAdmin && !req.user.isTournamentAdmin))
+    return next({ name: 'DomainActionError', error: 'notEnoughPermissions' }, req, res, null);
+
   try {
     await SchemeService.drawBracket(req.scheme);
     return res.json({});
@@ -65,6 +78,9 @@ const drawBracket = async (req, res, next) => {
 }
 
 const getScore = async (req, res, next) => {
+  if (!req.user || (!req.user.isAdmin && !req.user.isTournamentAdmin))
+    return next({ name: 'DomainActionError', error: 'notEnoughPermissions' }, req, res, null);
+
   try {
     const rankings = await SchemeService.getScore(req.scheme);
     return res.json(rankings);
@@ -75,6 +91,9 @@ const getScore = async (req, res, next) => {
 }
 
 const saveScore = async (req, res, next) => {
+  if (!req.user || (!req.user.isAdmin && !req.user.isTournamentAdmin))
+    return next({ name: 'DomainActionError', error: 'notEnoughPermissions' }, req, res, null);
+
   try {
     await SchemeService.saveScore(req.scheme, req.body);
     return res.json({});
@@ -85,15 +104,15 @@ const saveScore = async (req, res, next) => {
 }
 
 
-router.get('/:id/scores', adminIdentity, include, getScore);
-router.post('/:id/scores/save', adminIdentity, include, saveScore);
-router.get('/:id/drawBracket', adminIdentity, include, drawBracket);
+router.get('/:id/scores', identity, include, getScore);
+router.post('/:id/scores/save', identity, include, saveScore);
+router.get('/:id/drawBracket', identity, include, drawBracket);
 router.use('/:id/enrollments', include, require('../enrollment/enrollment.controller'));
 router.use('/:id/matches', include, require('../match/match.controller'));
 router.use('/:id/groups', include, require('../group/group.controller'));
 router.get('/:id', get);
-router.post('/:id', adminIdentity, update);
-router.post('/', adminIdentity, create);
-router.delete('/:id', adminIdentity, remove);
+router.post('/:id', identity, update);
+router.post('/', identity, create);
+router.delete('/:id', identity, remove);
 
 module.exports = router;
